Use linear interpolation for automatic corrections of short time span
This special case was implemented previously, before the slopes
surrounding level-shifts were accounted for by the automatic procedure,
at which point it provided significant benefit by producing 'smoother'
corrections.

The introduction of slope estimations often produces smoother
corrections on its own, and so the benefit of the reintroduction of this
special case is diminished. Yet, the special case still retains some
utility, for example, in the correction of nonsensical artifacts with
abberations small enough to be tolerated by the noise threshold, yet
still large enough to disturb the corrected trace; or when the slope
cannot be estimated.